refactor: move component converter interface to converter module
