+++ bricks-assemble/src/main/java/com/github/ingogriebsch/bricks/assemble/reader/common/SimpleComponentReader.java
@@ -21,10 +21,10 @@ package com.github.ingogriebsch.bricks.assemble.reader.common;
 
 import java.io.InputStream;
 
-import com.github.ingogriebsch.bricks.assemble.converter.ComponentConverter;
 import com.github.ingogriebsch.bricks.assemble.loader.ResourceLoader;
 import com.github.ingogriebsch.bricks.assemble.reader.ComponentReader;
 import com.github.ingogriebsch.bricks.assemble.utils.validate.Validator;
+import com.github.ingogriebsch.bricks.converter.ComponentConverter;
 import com.github.ingogriebsch.bricks.model.Component;
 
 import lombok.NonNull;
